requirements.txt
google-cloud-speech==1.3.2
google-cloud-storage==1.27.0
youtube-dl==2020.3.24
